+++ com.unity.mobile.android-logcat/Editor/AndroidLogcatStacktraceWindow.cs
@@ -15,7 +15,7 @@ namespace Unity.Android.Logcat
 #if PLATFORM_ANDROID
         static readonly string m_RedColor = "#ff0000ff";
         static readonly string m_GreenColor = "#00ff00ff";
-        static readonly string m_DefaultAddressRegex = @"\s*#\d{2}\s*pc\s([a-fA-F0-9]{8}).*(lib.*\.so)";
+        internal static readonly string m_DefaultAddressRegex = @"\s*#\d{2}\s*pc\s*(\S*)\s*.*(lib.*\.so)";
 
         enum WindowMode
         {
@@ -62,20 +62,6 @@ namespace Unity.Android.Logcat
             return false;
         }
 
-        string GetSymbolFile(string symbolPath, string libraryFile)
-        {
-            var fullPath = Path.Combine(symbolPath, libraryFile);
-            if (File.Exists(fullPath))
-                return fullPath;
-
-            // Try sym.so extension
-            fullPath = Path.Combine(symbolPath, Path.GetFileNameWithoutExtension(libraryFile) + ".sym.so");
-            if (File.Exists(fullPath))
-                return fullPath;
-
-            return null;
-        }
-
         void AddSymbolPath(string path)
         {
             int index = m_RecentSymbolPaths.IndexOf(path);
@@ -115,7 +101,7 @@ namespace Unity.Android.Logcat
                 else
                 {
                     string resolved = string.Format(" <color={0}>(Not resolved)</color>", m_RedColor);
-                    var symbolFile = GetSymbolFile(symbolPath, library);
+                    var symbolFile = AndroidLogcatUtilities.GetSymbolFile(symbolPath, library);
                     if (string.IsNullOrEmpty(symbolFile))
                     {
                         resolved = string.Format(" <color={0}>({1} not found)</color>", m_RedColor, library);
@@ -124,7 +110,7 @@ namespace Unity.Android.Logcat
                     {
                         try
                         {
-                            var result = Addr2LineWrapper.Run("\"" + symbolFile + "\"", new[] { address });
+                            var result = AndroidLogcatManager.instance.Runtime.Tools.RunAddr2Line(symbolFile, new[] { address });
                             AndroidLogcatInternalLog.Log("addr2line \"{0}\" {1}", symbolFile, address);
                             if (!string.IsNullOrEmpty(result[0]))
                                 resolved = string.Format(" <color={0}>({1})</color>", m_GreenColor, result[0].Trim());
@@ -265,7 +251,8 @@ namespace Unity.Android.Logcat
                 case WindowMode.ResolvedLog:
                     // Note: Not using EditorGUILayout.SelectableLabel, because scrollbars are not working correctly
                     EditorGUILayout.TextArea(m_ResolvedStacktraces, AndroidLogcatStyles.stacktraceStyle, GUILayout.ExpandHeight(true));
-                    GUIUtility.keyboardControl = 0;
+                    // Keep this commented, otherwise, it's not possible to select text in this text area and copy it.
+                    //GUIUtility.keyboardControl = 0;
                     break;
                 case WindowMode.OriginalLog:
                     m_Text = EditorGUILayout.TextArea(m_Text, AndroidLogcatStyles.stacktraceStyle, GUILayout.ExpandHeight(true));