+++ com.unity.mobile.android-logcat/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.4-preview] - 2020-02-29
+
+### Fix & Improvements.
+ - Added feature 'Filter by process id'
+ - Fixed addr2line functionality, when we try to resolve stacktrace
+ - Correctly open Terminal on macOS Catalina
+
 ## [1.0.3-preview] - 2019-11-17
 
 ### Fix & Improvements.